Separate client class added
